implement resource packs
This commit adds a ResourcePack class and changes the main method to
load a resource pack instead of having hard coded IDs and string paths.
I also changed the image replacing algorithm to account for the fact
that some character select icons aren't squares. This commit also
changes the Swapper to allow for swapping in arbitrary ByteArrays, which
is needed for reading the Zip file entries.